requirements.txt
pillow
opencv-python
numpy
rembg